agent/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
agno==1.0.1
openai>=1.30.0
pydantic==2.8.0
asyncpg==0.29.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
sqlalchemy
httpx>=0.27.0
PyJWT[crypto]>=2.8.0
